exception handler and exceptions created

+++ src/main/java/com/racedriverlife/racedriverlife_app/services/RaceService.java
@@ -1,14 +1,17 @@
 package com.racedriverlife.racedriverlife_app.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.racedriverlife.racedriverlife_app.entities.Race;
 import com.racedriverlife.racedriverlife_app.repositories.RaceRepository;
+import com.racedriverlife.racedriverlife_app.services.exceptions.DatabaseException;
+import com.racedriverlife.racedriverlife_app.services.exceptions.ResourceNotFoundException;
 
-import jakarta.transaction.Transactional;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class RaceService {
@@ -21,21 +24,34 @@ public class RaceService {
 	}
 	
 	public Race getRaceById(Long id) {
-		return this.repository.findById(id).get();
+		Optional<Race> obj = repository.findById(id);
+		return obj.orElseThrow( () -> new ResourceNotFoundException(id) );
 	}
 	
 	public Race save(Race race) {
 		return this.repository.save(race);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+
 	
 	public Race update(Long id, Race race) {
-		Race entity = repository.getReferenceById(id);
-		entity = updateData(entity, race);
-		return repository.save(entity);
+		try {
+			Race entity = repository.getReferenceById(id);
+			entity = updateData(entity, race);
+			return repository.save(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+	}
+	
+	public void delete(Long id) {
+		if (repository.existsById(id)) {
+			repository.deleteById(id);
+		}
+		else {
+			throw new DatabaseException("Resource not found. Id " + id);
+		}
 	}
 		
 	private Race updateData(Race entity, Race race) {
