+++ src/main/java/com/racedriverlife/racedriverlife_app/services/TaskService.java
@@ -1,6 +1,8 @@
 package com.racedriverlife.racedriverlife_app.services;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,8 +13,11 @@ import com.racedriverlife.racedriverlife_app.entities.Task;
 import com.racedriverlife.racedriverlife_app.entities.enums.TaskStatus;
 import com.racedriverlife.racedriverlife_app.repositories.RaceRepository;
 import com.racedriverlife.racedriverlife_app.repositories.TaskRepository;
+import com.racedriverlife.racedriverlife_app.services.exceptions.DatabaseException;
+import com.racedriverlife.racedriverlife_app.services.exceptions.RaceNotFoundException;
+import com.racedriverlife.racedriverlife_app.services.exceptions.ResourceNotFoundException;
 
-import jakarta.transaction.Transactional;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class TaskService {
@@ -31,38 +36,57 @@ public class TaskService {
 	}
 	
 	public Task getTaskById(Long id) {
-		return this.repository.findById(id).get();
+		Optional<Task> obj = repository.findById(id);
+		return obj.orElseThrow( () -> new ResourceNotFoundException(id) ); 
 	}
 		
 	public Task save(TaskDTO taskDTO) {		
-		Task convertedTask = convertDTOtoEntity(taskDTO);
-		
-		this.repository.save(convertedTask);
-		
-		Race raceUpdated = updateRaceData(taskDTO.getRaceId());
-		
-		raceService.update(taskDTO.getRaceId(), raceUpdated);
-		
-		return convertedTask;
+		try {
+			Task convertedTask = convertDTOtoEntity(taskDTO);
+			
+			this.repository.save(convertedTask);
+			
+			Race raceUpdated = updateRaceData(taskDTO.getRaceId());
+			
+			raceService.update(taskDTO.getRaceId(), raceUpdated);
+			
+			return convertedTask;
+		}
+		catch (NoSuchElementException e) {
+			throw new RaceNotFoundException(taskDTO.getRaceId());
+		}
 	}
 	
 
-	public void delete(Long id) {
-		Task task = this.getTaskById(id);
-		
-		Long raceId = task.getRace().getRaceId();
-		
-		repository.deleteById(id);
-		
-		updateRaceData(raceId);
-	}
+
 	
 	public Task update(Long id, TaskDTO taskDTO) { 
-		Task entity = repository.getReferenceById(id);
 		
-		entity = updateData(entity, taskDTO);
 		
-		return this.repository.save(entity);
+		try {
+			Task entity = repository.getReferenceById(id);
+			entity = updateData(entity, taskDTO);
+			return this.repository.save(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (NoSuchElementException e) {
+			throw new ResourceNotFoundException(id);
+		}
+	}
+	
+	public void delete(Long id) {
+		if (repository.existsById(id)) {
+			Task task = this.getTaskById(id);
+			
+			Long raceId = task.getRace().getRaceId();
+			
+			repository.deleteById(id);
+			
+			updateRaceData(raceId);
+		}
+		else {
+			throw new DatabaseException("Resource not found. Id " + id);
+		}
 	}
 	
 	private Task updateData(Task entity, TaskDTO task) {
@@ -94,11 +118,15 @@ public class TaskService {
 	
 	
 	public Race updateRaceData(Long id) {
-		Race savedRace = raceRepository.findById(id).get();
-		
-		savedRace.countTotalTasks();
-		
-		System.out.println("Corrida atualizada");
-		return raceService.update(id, savedRace);
+		try {
+			Race savedRace = raceRepository.findById(id).get();
+			
+			savedRace.countTotalTasks();
+			
+			System.out.println("Corrida atualizada");
+			return raceService.update(id, savedRace);
+		} catch (NoSuchElementException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 }
