+++ src/main/java/com/racedriverlife/racedriverlife_app/services/UserService.java
@@ -2,8 +2,10 @@ package com.racedriverlife.racedriverlife_app.services;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.racedriverlife.racedriverlife_app.DTOs.UserDTO;
@@ -11,6 +13,11 @@ import com.racedriverlife.racedriverlife_app.entities.Race;
 import com.racedriverlife.racedriverlife_app.entities.RaceCentral;
 import com.racedriverlife.racedriverlife_app.entities.User;
 import com.racedriverlife.racedriverlife_app.repositories.UserRepository;
+import com.racedriverlife.racedriverlife_app.services.exceptions.DatabaseException;
+import com.racedriverlife.racedriverlife_app.services.exceptions.ResourceNotFoundException;
+import com.racedriverlife.racedriverlife_app.services.exceptions.UserAlreadyExistsException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -25,19 +32,30 @@ public class UserService {
 		for (User x : user) {
 			userDTO.add(convertToDTO(x));
 		}
+		
 		return userDTO;
 		
 	}
 
 	public User getUserById(Long id) {
-		return this.repository.findById(id).get();
+		Optional<User> obj = repository.findById(id);
+		return obj.orElseThrow( () -> new ResourceNotFoundException(id) );
 	}
 	
 	public User save(UserDTO userDTO) {
+		List<UserDTO> allUsers = this.getAllUsers();
+		
+		for (UserDTO user : allUsers) {
+			if (user.getUserName().equals(userDTO.getUserName())) {
+				throw new UserAlreadyExistsException();
+			}
+		}
 		
 		User user = new User();
 		user.setUserName(userDTO.getUserName());
 		user.setPassword(userDTO.getPassword());
+		
+		
 		
 		RaceCentral raceCentral = new RaceCentral();
 		raceCentral.setRacesDisputed(0);
@@ -55,16 +73,40 @@ public class UserService {
 		return this.repository.save(user);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+
 	
 	public User update(Long id, UserDTO userDTO) {
-		User entity = repository.getReferenceById(id); // pega todos os atributos do user especifico
+
 		
-		updateData(entity, userDTO);
+		try {
+			User entity = repository.getReferenceById(id); // pega todos os atributos do user especifico
+			
+			if (!userDTO.getUserName().equals(entity.getUserName())) {
+				List<UserDTO> allUsers = this.getAllUsers();
+				
+				for (UserDTO user : allUsers) {
+					if (user.getUserName().equals(userDTO.getUserName())) {
+						throw new UserAlreadyExistsException();
+					}
+				}
+			}
+			
+			updateData(entity, userDTO);
+			return repository.save(entity);	
+		}
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+	}
+	
+	public void delete(Long id) {
 		
-		return repository.save(entity);	
+		if (repository.existsById(id)) {
+			repository.deleteById(id);
+		}
+		else {
+			throw new DatabaseException("Resource not found. Id " + id);
+		}
 	}
 	
 	private void updateData(User entity, UserDTO userDTO) { // atualizando só o atualizável
